+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added Action Manager to the component menu
+- Added a *simple node view* option in the tool menu. Simple node view displays the tree how it was, turning this off will show the inspector view inside each node and removes the inspector from the side
+
+### Changed
+
+- Changed the title of the visual editor when using a custom version to "Decision Tree Editor (Custom)"
 
 ## 1.0.5